Organize source code and provide project README
