+++ scripts/generate_output
@@ -1,0 +1,5 @@
+#!/usr/bin/env python
+from creating_mv.generate_output import main
+
+if __name__ == "__main__":
+    main()
