+++ scripts/preprocess_data
@@ -1,0 +1,5 @@
+#!/usr/bin/env python
+from creating_mv.preprocess_data import main
+
+if __name__ == "__main__":
+    main()
