+++ scripts/train_model
@@ -1,0 +1,5 @@
+#!/usr/bin/env python
+from creating_mv.train_model import main
+
+if __name__ == "__main__":
+    main()
